requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Excel file processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AWS integration
boto3>=1.34.0
botocore>=1.34.0

# Progress bars and utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring and logging
structlog>=23.2.0
prometheus-client>=0.17.0

# Data validation
cerberus>=1.3.0

# Date/time handling
python-dateutil>=2.8.0

# HTTP client
httpx>=0.25.0

# JSON handling
orjson>=3.9.0

# Configuration management
pydantic-settings>=2.0.0 